+++ modules/core/src/ClassicRegister.cpp
@@ -7,7 +7,7 @@ std::string ClassicRegister::getName() const{
     return name;
 }
 
-void ClassicRegister::setSize(const size_t& size){
+void ClassicRegister::setSize(const int& size){
     this->size = size;
     this->data.clear();
 }
@@ -24,7 +24,7 @@ Bit ClassicRegister::operator[] (int position) const {
     throw InvalidInputException("Invalid position");
 }
 
-bool ClassicRegister::linkRegister(size_t bitPos, size_t quantumPos, const std::string& link, std::string& errorMsg){
+bool ClassicRegister::linkRegister(int bitPos, int quantumPos, const std::string& link, std::string& errorMsg){
     errorMsg = "";
     for(const Bit& bit : data){
         if(bit.bitPos == bitPos && bit.link == link){
